(svn r22598) [1.1] -Backport from trunk: - Fix: [NoAI] Doing rescan_ai in a game with running AIs caused a crash [FS#4631] (r22534) - Fix: Do not create an implicit order if the current order is the first order in the order list and we visit the station of the last entry of the order list (r22532) - Fix: MinGW 64 related compilation issues [FS#4623] (r22522, r22491, r22490, r22489) - Fix: The layout selectors of the airport build GUI did not latch properly (r22497, r22495)

+++ src/table/settings.h
@@ -34,7 +34,7 @@ static bool DifficultyChange(int32);
 static bool DifficultyNoiseChange(int32 i);
 static bool MaxNoAIsChange(int32 i);
 static bool CheckRoadSide(int p1);
-static int32 ConvertLandscape(const char *value);
+static size_t ConvertLandscape(const char *value);
 static bool CheckFreeformEdges(int32 p1);
 static bool ChangeDynamicEngines(int32 p1);
 static bool StationCatchmentChanged(int32 p1);
